S002-T10 - correction after review part 4
