+++ src/main/java/com/project/stockexchangeappbackend/entity/OrderType.java
@@ -1,0 +1,6 @@
+package com.project.stockexchangeappbackend.entity;
+
+public enum OrderType {
+    BUYING_ORDER,
+    SELLING_ORDER
+}
