+++ src/main/java/com/project/stockexchangeappbackend/entity/PriceType.java
@@ -1,0 +1,7 @@
+package com.project.stockexchangeappbackend.entity;
+
+public enum PriceType {
+    EQUAL,
+    GREATER_OR_EQUAL,
+    LESS_OR_EQUAL
+}
