+++ src/main/java/com/project/stockexchangeappbackend/entity/Resource.java
@@ -17,11 +17,11 @@ public class Resource {
     @GeneratedValue(generator = "RESOURCE_SEQUENCE")
     private Long id;
 
-    @ManyToOne(targetEntity = User.class, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @ManyToOne(targetEntity = User.class)
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-    @ManyToOne(targetEntity = Stock.class, cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @ManyToOne(targetEntity = Stock.class)
     @JoinColumn(name = "stock_id", nullable = false)
     private Stock stock;
